backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosmtplib==2.0.2
albucore==0.0.24
albumentations==2.0.6
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.0.1
blinker==1.8.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cnocr==2.3.1
cnstd==1.2.5.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
doclayout_yolo==0.0.3
ecdsa==0.19.0
email_validator==2.2.0
evaluate==0.4.3
fastapi==0.112.2
fastapi-cli==0.0.7
fastapi-mail==1.4.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
latex2mathml==3.78.0
lightning-utilities==0.14.3
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
onnx==1.17.0
onnxruntime==1.21.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optimum==1.24.0
orjson==3.10.15
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pix2text @ git+https://github.com/breezedeus/pix2text.git@1c1aecac4e9041242232059709447c743dc61704
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.2
pydantic-settings==2.4.0
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.25.5
pyngrok==7.2.7
pyparsing==3.2.3
pyreadline3==3.5.4
pyspellchecker==0.8.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
rapidocr-onnxruntime==1.3.25
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
safetensors==0.5.3
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools==80.3.1
shapely==2.1.0
shellingham==1.5.4
simsimd==6.2.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
stringzilla==3.12.5
sympy==1.14.0
thop==0.1.1.post2209072238
tokenizers==0.21.1
torch==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
transformers==4.48.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.130
ultralytics-thop==2.0.14
Unidecode==1.4.0
urllib3==2.3.0
uuid7==0.1.0
uvicorn==0.30.6
wandb==0.19.11
watchfiles==1.0.4
websockets==15.0
xxhash==3.5.0
yarl==1.20.0
